+++ Plugins/AutoPlayGifs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # AutoPlayGifs Changelog
 
+### 0.0.2
+
+ - Remove stuff discord added
+ - Add member list avatars
+
 ### 0.0.1
 
  - Initial release